cleanup

+++ src/Polyfill/Ensure/Ensure_NotNegative.cs
@@ -1,5 +1,6 @@
 namespace Polyfills;
 
+// ReSharper disable once RedundantUsingDirective
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using System;
@@ -46,7 +47,7 @@ static partial class Ensure
     /// <param name="name">The name of the parameter with which <paramref name="value"/> corresponds.</param>
     public static nint NotNegative(nint value, [CallerArgumentExpression(nameof(value))] string? name = null)
     {
-        if (value < (nint) 0)
+        if (value < 0)
         {
             ThrowNegative(value, name);
         }
