+++ src/Polyfill/Ensure/Ensure_NotNegativeOrZero.cs
@@ -1,9 +1,10 @@
 namespace Polyfills;
 
-using System.Numerics;
-using System.Runtime.CompilerServices;
 using System;
 using System.Diagnostics.CodeAnalysis;
+// ReSharper disable once RedundantUsingDirective
+using System.Numerics;
+using System.Runtime.CompilerServices;
 
 #if PolyPublic
 public
@@ -46,7 +47,7 @@ static partial class Ensure
     /// <param name="name">The name of the parameter with which <paramref name="value"/> corresponds.</param>
     public static nint NotNegativeOrZero(nint value, [CallerArgumentExpression(nameof(value))] string? name = null)
     {
-        if (value <= (nint) 0)
+        if (value <= 0)
         {
             ThrowNegativeOrZero(value, name);
         }
